feat: Provider new structure, create transaction
